Move `task_group_to_dict` functions from task-sdk to server-side & reorganize tests
Follow-up of the fix in https://github.com/apache/airflow/pull/54756
- Move `task_group_to_dict` and `task_group_to_dict_grid` functions from `task-sdk` to `airflow-core` API services
- Update import paths in `grid.py` and `structure.py` to use new server-side location
- Update deprecation mappings to point to new FastAPI service module locations
- Remove `AbstractOperator` from isinstance checks in server-side functions (now handles only serialized objects)
- Split mixed tests: moved TaskGroup functionality tests to `task-sdk`, keep server function tests in `airflow-core`
- Add comprehensive TaskGroup tests in `task-sdk` covering creation, relationships, decorators, and validation
- Create clear architectural boundary: client-side TaskGroup authoring vs server-side serialized DAG processing
