+++ airflow-core/src/airflow/api_fastapi/core_api/routes/ui/structure.py
@@ -30,9 +30,9 @@ from airflow.api_fastapi.core_api.services.ui.structure import (
     bind_output_assets_to_tasks,
     get_upstream_assets,
 )
+from airflow.api_fastapi.core_api.services.ui.task_group import task_group_to_dict
 from airflow.models.dag_version import DagVersion
 from airflow.models.serialized_dag import SerializedDagModel
-from airflow.sdk.definitions.taskgroup import task_group_to_dict
 from airflow.utils.dag_edges import dag_edges
 
 structure_router = AirflowRouter(tags=["Structure"], prefix="/structure")
